api key was done

+++ src/https/routes/delete.rs
@@ -1,7 +1,9 @@
 use std::fs;
 use rocket_contrib::json::Json;
 use serde::Serialize;
-use rocket::get;
+use rocket::{get, State};
+
+use crate::https::StateData;
 
 
 #[derive(Serialize)]
@@ -9,8 +11,13 @@ pub struct Data {
     error: String,
 }
 
-#[get("/delete/<path>")]
-pub fn delete(mut path: String) -> Json<Data> {
+#[get("/delete/<path>/<api_key>")]
+pub fn delete(mut path: String, api_key: String, api_state : State<StateData>) -> Json<Data> {
+    if api_key != api_state.api_key{
+        return Json(Data{
+            error : "Not authorized".to_string()
+        })
+    }
     path = path.replace("`", "/");
     path = "database/".to_string() + &*path + ".txt";
     let delete_error = fs::remove_file(path);
